+++ tests/test_reference.py
@@ -4,6 +4,11 @@
 from pyiem import reference
 
 
+def test_ugc_state_names_in_state_names():
+    """Test that ugc_state_names is a superset of state_names."""
+    assert all(x in reference.ugc_state_names for x in reference.state_names)
+
+
 def test_states():
     """Test that we have the same number of states"""
     assert len(reference.state_names) == len(reference.state_bounds)