+++ CHANGELOG.md
@@ -26,6 +26,8 @@ All notable changes to this library are documented in this file.
 - Account for FastAPI usage of `webutil.ListOrCSVType`.
 - Add `ip_throttle_secs` to `webutil.iemapp` to deal with IEM pain.
 - Add knob to `ugcs_to_text` to control total generated message size.
+- Add reference `ugc_state_names` to provide the two character prefix codes
+  used within NWS UGCs.
 - Gracefully handle `XTEUS` product without a value set.
 - Improve `iemapp` to better capture actual HTTP status_code and document
   what happens during Exception to status_code mapping.